[Fix] 필터 쓰는 항목 최적화
- useGetFilter 훅 제작
- 해당 훅을 Filter 및 createClub2 동시에 사용
