+++ src/pages/CreateClub/CreateClub2/CreateClub2.jsx
@@ -1,16 +1,17 @@
-import { MainButton, NomalTitle } from '@/components/Atoms';
+import { GenreButton, MainButton, NomalTitle } from '@/components/Atoms';
+import useGetFilter from '@/pages/Filter/useGetFilter';
 import useCreateClubStore from '@/store/useCreateClubStore';
 import { getDocumentTitle } from '@/utils';
 import { Helmet } from 'react-helmet-async';
-import { useLoaderData } from 'react-router-dom';
 
 export function CreateClub2() {
-  const genres = useLoaderData();
   const { clubInfo, addGenre, removeGenre } = useCreateClubStore((state) => ({
     clubInfo: state.clubInfo,
     addGenre: state.addGenre,
     removeGenre: state.removeGenre,
   }));
+
+  const filterData = useGetFilter();
 
   const handlegenreButton = (e) => {
     e.preventDefault();
@@ -42,11 +43,16 @@ export function CreateClub2() {
               책 장르 (1개 필수 선택)
             </span>
             <ul className="flex flex-wrap gap-3 py-4">
-              <GenreButton
-                genres={genres}
-                state={clubInfo}
-                onClick={handlegenreButton}
-              />
+              {filterData.map((filterData) => {
+                return (
+                  <GenreButton
+                    key={filterData.id}
+                    filterData={filterData}
+                    state={clubInfo}
+                    onClick={handlegenreButton}
+                  />
+                );
+              })}
             </ul>
           </div>
         </div>
@@ -62,25 +68,4 @@ export function CreateClub2() {
       </main>
     </>
   );
-}
-
-function GenreButton({ className, state, genres, onClick, ...restProps }) {
-  const BASE_SYTLE = {
-    className: 'h-9 rounded-5xl border px-4 text-b-2-medium',
-  };
-
-  return genres.map(({ id, title }) => (
-    <li key={id}>
-      <button
-        type="button"
-        name={title}
-        value={id}
-        onClick={onClick}
-        className={`${BASE_SYTLE.className} ${className} ${state.genre !== null && state.genre.includes(id) ? 'bg-black text-white' : 'bg-white text-bjblack'}`}
-        {...restProps}
-      >
-        {title}
-      </button>
-    </li>
-  ));
 }
