--- src/pages/CreateClub/CreateClub2/loader.js
@@ -1,6 +1,0 @@
-import pb from '@/api/pocketbase';
-
-export async function loader() {
-  const data = await pb.collection('genres').getFullList();
-  return data;
-}
